Various changes related to deferred updates:
Add functions to update the UI immediately (either for a specific node or for all nodes).
Changed ko.evaluateAsynchronously to just use setTimeout.
Fixed bug that a computed observable could be evaluated after it was disposed.
Re-arrange throttle extender slightly: it will do either write throttling or evaluation throttling but not both.
Turn off deferred updates for a computed observable during the write so that the 'write' is registered immediately (assuming that the read function accesses any observables that are written to).