Finally have working light-weight `with` binding.
If the initial value is non-false, it'll save a copy of the nodes and bind the original nodes.
If the value changes (but doesn't become false), it will simply update the binding without removing or replacing nodes.
If the value changes to false, it will remove the nodes from the document.
If the value changes to non-false, it will copy the saved nodes into the document and bind the copied nodes.